feat : RowMapper 인터페이스 대신에 구현체로 변경
