+++ jdbc/src/main/java/camp/nextstep/jdbc/core/RowMapper.java
@@ -1,9 +1,38 @@
 package camp.nextstep.jdbc.core;
 
+import java.lang.reflect.Field;
+import java.lang.reflect.InvocationTargetException;
 import java.sql.ResultSet;
-import java.sql.SQLException;
+import java.util.Arrays;
 
-@FunctionalInterface
-public interface RowMapper<T> {
-    T mapRow(final ResultSet resultSet) throws SQLException;
+public class RowMapper<T> {
+
+    private final Class<T> clazz;
+
+    public RowMapper(Class<T> clazz) {
+        this.clazz = clazz;
+    }
+
+    public T mapRow(ResultSet resultSet) {
+        try {
+            T instance = clazz.getDeclaredConstructor().newInstance();
+            setFieldValues(instance, clazz.getDeclaredFields(), resultSet);
+            return instance;
+        } catch (NoSuchMethodException | InstantiationException | IllegalAccessException |
+                 InvocationTargetException e) {
+            throw new RuntimeException(e);
+        }
+    }
+
+    private void setFieldValues(Object instance, Field[] fields, ResultSet resultSet) {
+        Arrays.stream(fields)
+            .forEach(field -> {
+                try {
+                    field.setAccessible(true);
+                    field.set(instance, ResultSetColumnReader.read(resultSet, field.getName(), field.getType()));
+                } catch (IllegalAccessException e) {
+                    throw new RuntimeException(e);
+                }
+            });
+    }
 }
